[Builder] Only build resources that reside in the currently build project, don't create contents folder if project does not contain xdoc resources

+++ org.eclipse.xtext.xdoc.ui/src/org/eclipse/xtext/xdoc/ui/builder/XdocBuilderParticipant.java
@@ -16,21 +16,20 @@ import org.eclipse.xpand2.output.VetoException;
 import org.eclipse.xtend.type.impl.java.JavaBeansMetaModel;
 import org.eclipse.xtext.builder.IXtextBuilderParticipant;
 import org.eclipse.xtext.resource.IResourceDescription;
+import org.eclipse.xtext.util.Wrapper;
 
 public class XdocBuilderParticipant implements IXtextBuilderParticipant {
 
 	public void build(final IBuildContext context, IProgressMonitor monitor)
 			throws CoreException {
-		final IFolder folder = context.getBuiltProject().getFolder("contents");
-		if (!folder.exists())
-			folder.create(true, true, monitor);
+		final Wrapper<IFolder> folder = Wrapper.wrap(null);
 
 		OutputImpl output = new OutputImpl();
 		Outlet outlet = new Outlet() {
 			@Override
 			public FileHandle createFileHandle(String fileName)
 					throws VetoException {
-				IFile file = folder.getFile(new Path(fileName));
+				IFile file = folder.get().getFile(new Path(fileName));
 				return new EclipseBasedFileHandle(file, this);
 			}
 		};
@@ -39,18 +38,28 @@ public class XdocBuilderParticipant implements IXtextBuilderParticipant {
 				null, null, null, null);
 		ctx.registerMetaModel(new JavaBeansMetaModel());
 
+		String projectName = context.getBuiltProject().getName();
 		for (IResourceDescription.Delta delta : context.getDeltas()) {
 			// handle deletion
-			if (delta.getUri().fileExtension().equals("xdoc")) {
-				if (delta.getNew() == null) {
-					IFile file = folder.getFile(delta.getUri().lastSegment()+".html");
-					if (file.exists())
-						file.delete(true, monitor);
-				} else {
-					Resource resource = context.getResourceSet().getResource(
-							delta.getUri(), true);
-					EObject object = resource.getContents().get(0);
-					generate(object, ctx, context);
+			String projectSegment = delta.getUri().segment(1);
+			if (projectName.equals(projectSegment)) {
+				if (delta.getUri().fileExtension().equals("xdoc")) {
+					if (folder.get() == null) {
+						IFolder contentsFolder = context.getBuiltProject().getFolder("contents");
+						if (!contentsFolder.exists())
+							contentsFolder.create(true, true, monitor);
+						folder.set(contentsFolder);
+					}
+					if (delta.getNew() == null) {
+						IFile file = folder.get().getFile(delta.getUri().lastSegment()+".html");
+						if (file.exists())
+							file.delete(true, monitor);
+					} else {
+						Resource resource = context.getResourceSet().getResource(
+								delta.getUri(), true);
+						EObject object = resource.getContents().get(0);
+						generate(object, ctx, context);
+					}
 				}
 			}
 		}
